// src/utils/interfaces.ts
export enum DelegationTypes {
  COMPOUND = "compound-governor",
  SPLIT_DELEGATION = "split-delegation"
}
// ...
export interface TreasuryWallet {
  name: string;
  address: string;
  network: string;
}
// ...
export interface ExtendedSpace {
  id: string;
  name: string;
  symbol: string;
  network: string;
  strategies: SpaceStrategy[];
  delegationPortal: DelegatesConfig;
  about: string;
  avatar: string;
  skin: string;
  domain: string | null;
  website: string | null;
  terms: string | null;
  coingecko: string | null;
  github: string | null;
  twitter: string | null;
  followersCount: number;
  private: boolean;
  admins: string[];
  moderators: string[];
  members: string[];
  categories: string[];
  parent: ExtendedSpace | null;
  children: ExtendedSpace[];
  filters: { minScore: number; onlyMembers: boolean };
  plugins: Record<string, any>;
  validation: SpaceValidation;
  voteValidation: VoteValidation;
  treasuries: TreasuryWallet[];
  template: string;
  guidelines: string;
  verified: boolean;
  flagged: boolean;
  hibernated: boolean;
  turbo: boolean;
  boost: {
    enabled: boolean;
    bribeEnabled: boolean;
  };
  voting: {
    delay: number | null;
    hideAbstain: boolean;
    period: number | null;
    quorum: number | null;
    quorumType: "default" | "rejection";
    type: string | null;
    privacy: string | null;
  };
}

export interface DelegatesConfig {
  delegationType: DelegationTypes;
  delegationContract: string;
  delegationNetwork: string;
  delegationApi: string;
}
export interface SpaceValidation {
  name: string;
  params: Record<string, any>;
}

export interface SpaceStrategy {
  name: string;
  network: string;
  params: Record<string, unknown>;
}

export interface ProposalSpace {
  id: string;
  name: string;
  members: string[];
  avatar: string;
  symbol: string;
  verified: boolean;
  turbo: boolean;
}

export interface Proposal {
  id: string;
  title: string;
  ipfs: string;
  network: string;
  choices: string[];
  type: string;
  snapshot: string;
  author: string;
  body: string;
  created: number;
  start: number;
  end: number;
  state: string;
  symbol: string;
  privacy: string;
  validation: VoteValidation;
  discussion: string;
  quorum: number;
  quorumType: "default" | "rejection";
  scores: number[];
  scores_state: string;
  scores_total: number;
  scores_by_strategy: number[][];
  votes: number;
  plugins: Record<string, any>;
  space: ExtendedSpace;
  strategies: SpaceStrategy[];
  flagged: boolean;
}

export interface VoteValidation {
  name: string;
  params: Record<string, any>;
}